+++ cratedb_toolkit/io/processor/kinesis_lambda.py
@@ -25,7 +25,7 @@ Resources:
 # /// script
 # requires-python = ">=3.9"
 # dependencies = [
-#   "commons-codec==0.0.3",
+#   "commons-codec==0.0.6",
 #   "sqlalchemy-cratedb==0.38.0",
 # ]
 # ///