+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Processor: Updated Kinesis Lambda processor to understand AWS DMS
 - MongoDB: Fix missing output on STDOUT for `migr8 export`
 - MongoDB: Improve timestamp parsing by using `python-dateutil`
+- MongoDB: Converge `_id` input field to `id` column instead of dropping it
 
 ## 2024/07/25 v0.0.16
 - `ctk load table`: Added support for MongoDB Change Streams