+++ CHANGES.md
@@ -13,6 +13,7 @@
 - MongoDB: Strip single leading underscore character from all top-level fields
 - MongoDB: Map OID types to CrateDB TEXT columns
 - MongoDB: Make `migr8 extract` and `migr8 export` accept the `--limit` option
+- MongoDB: Fix indentation in prettified SQL output of `migr8 translate`
 
 ## 2024/07/25 v0.0.16
 - `ctk load table`: Added support for MongoDB Change Streams